Move monthly contribution related models into their own package
