+++ src/main/scala/com/gu/support/workers/model/monthlyContributions/state/FailureHandlerState.scala
@@ -1,8 +1,9 @@
-package com.gu.support.workers.model.state
+package com.gu.support.workers.model.monthlyContributions.state
 
 import java.util.UUID
 
 import com.gu.support.workers.model._
+import com.gu.support.workers.model.monthlyContributions.Contribution
 
 case class FailureHandlerState(
   requestId: UUID,
