+++ src/main/scala/com/gu/support/workers/model/monthlyContributions/state/StepFunctionUserState.scala
@@ -1,4 +1,4 @@
-package com.gu.support.workers.model.state
+package com.gu.support.workers.model.monthlyContributions.state
 
 import com.gu.support.workers.model.User
