--- src/main/scala/com/gu/support/workers/model/state/CreateSalesforceContactState.scala
@@ -1,12 +1,0 @@
-package com.gu.support.workers.model.state
-
-import java.util.UUID
-
-import com.gu.support.workers.model.{Contribution, PaymentMethod, User}
-
-case class CreateSalesforceContactState(
-  requestId: UUID,
-  user: User,
-  contribution: Contribution,
-  paymentMethod: PaymentMethod
-) extends StepFunctionUserState
